% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential-sections.R
\name{begin_section}
\alias{begin_section}
\title{Begin a section.}
\usage{
begin_section(section, container_class, title_container = h2)
}
\arguments{
\item{section}{section object.}

\item{container_class}{additional HTML class attribute}

\item{container}{a function to generate an HTML element to contain the section title.}
}
\value{
section start elements.
}
\description{
This renders the section title.
}
