% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_clicker.R
\name{clicker_question}
\alias{clicker_question}
\title{Create a clicker question.}
\usage{
clicker_question(
  title,
  ...,
  body = NULL,
  in_class_only = FALSE,
  container_class = NULL,
  force_render = FALSE,
  btn_text = "Show Clicker Question"
)
}
\arguments{
\item{title}{question title.}

\item{...}{5 answer options. They will be numbered 'A'-'E'.}

\item{body}{optional question body.}

\item{in_class_only}{if the question should only be shown 'in-class' or also when printing out the slides.}

\item{container_class}{additional HTML classes to add to the container.}

\item{force_render}{should the question always be rendered?}

\item{btn_text}{text for the "show" button.}
}
\description{
By default, clicker questions are only rendered if the option \code{stat305templates.render_clicker_questions} is
\code{TRUE} or if it is a function which returns \code{TRUE}.
}
