% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential-sections.R
\name{create_sections}
\alias{create_sections}
\title{Create a series of sections, displayed sequentially.}
\usage{
create_sections(..., .group_id)
}
\arguments{
\item{...}{named sections, to be displayed in the given order.}

\item{.group_id}{id for this group of sections.}
}
\value{
a list of sections.
}
\description{
To begin a section, call \code{\link[=begin_section]{begin_section()}} with the desired section, to end a section, call \code{\link[=end_section]{end_section()}}.
}
