% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential-sections.R
\name{end_section}
\alias{end_section}
\title{End a section.}
\usage{
end_section(
  label = "Submit Question",
  confirm_dialog,
  success_dialog,
  error_dialog,
  btn_class = "btn-primary btn-lg",
  container = div
)
}
\arguments{
\item{label}{submission button text.}

\item{confirm_dialog}{a modal dialog created with \code{\link[shiny:modalDialog]{shiny::modalDialog()}} to display for confirming submission of the
section. If \code{NULL}, or \code{FALSE}, no confirmation will be shown. This dialog needs to have at least one button
with the \code{data-exam} attribute set to \code{continue}!}

\item{success_dialog}{a modal dialog created with \code{\link[shiny:modalDialog]{shiny::modalDialog()}} to display after successful submission
of the section. If \code{NULL}, or \code{FALSE}, no dialog will be shown.}

\item{error_dialog}{a modal dialog created with \code{\link[shiny:modalDialog]{shiny::modalDialog()}} to display after successful submission
of the section. If \code{NULL}, or \code{FALSE}, no dialog will be shown. This dialog needs to have at least one button
with the \code{data-exam} attribute set to \code{reload}!}

\item{container}{a function to generate an HTML element to contain the button.}

\item{section}{section object.}
}
\value{
section end elements.
}
\description{
This renders a button to submit the answers to the section.
When users click the button, a modal dialog is shown for the user to confirm they want to submit the section.
}
