% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_safe.R
\name{get_safe_evaluator}
\alias{get_safe_evaluator}
\title{Get a function to safely evaluate user code.}
\usage{
get_safe_evaluator(priority, envir, rlimits, allow_env)
}
\arguments{
\item{priority, rlimits, allow_env}{see \link{enable_safe_eval} for details.}
}
\value{
a \verb{function(expr, timeout)} which evaluates the given expression \code{expr} safely in at most \code{timeout} seconds.
}
\description{
Get a function to safely evaluate user code.
}
