% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_session_data}
\alias{get_session_data}
\title{Get the default session data}
\usage{
get_session_data(
  what,
  fallback = NULL,
  session = getDefaultReactiveDomain(),
  asis = FALSE
)
}
\description{
Get the default session data
}
