% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential-sections.R
\name{init_sections}
\alias{init_sections}
\title{Initialize sections for this session}
\usage{
init_sections(
  sections,
  hide = TRUE,
  randomize_order = FALSE,
  session = getDefaultReactiveDomain()
)
}
\arguments{
\item{sections}{the sections object}

\item{hide}{should sections be hidden by default.}

\item{randomize_order}{randomize the order of sections. Can be a vector defining "strata"
(\code{NA} meaning "leave section in this order").}
}
\value{
sections in the new order.
}
\description{
Initialize sections for this session
}
