% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_clicker.R
\name{knit_print.clicker_question}
\alias{knit_print.clicker_question}
\title{Knitr clicker question print methods}
\usage{
\method{knit_print}{clicker_question}(x, ...)
}
\arguments{
\item{x}{An R object to be printed}

\item{...}{Additional arguments passed to the S3 method. Currently ignored,
except two optional arguments \code{options} and \code{inline}; see
the references below.}
}
\description{
\link[knitr:knit_print]{knitr::knit_print} method for \link{clicker_question}.
}
