% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rmd.R
\name{precompile_rmd}
\alias{precompile_rmd}
\title{Precompile HTML files}
\usage{
precompile_rmd(
  file,
  data = list(),
  runtime = "static",
  quiet = TRUE,
  precompiled_path,
  hasher = digest,
  section_divs = FALSE,
  md_extensions = "+tex_math_dollars",
  df_print = "kable",
  ...
)
}
\arguments{
\item{file}{path to the RMarkdown file}

\item{data}{a list of data to be made available to the Rmd file during rendering. Quoted items will be evaluated
on the server before knitting the Rmd file. Quoted expressions are evaluated in the \strong{server context}.}

\item{runtime, quiet}{arguments passed on to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}

\item{precompiled_path}{path where precompiled HTML file should be saved.}

\item{hasher}{function to compute the unique hash for the input data.}

\item{section_divs, md_extensions, df_print...}{passed on to \code{\link[rmarkdown:html_fragment]{rmarkdown::html_fragment()}}.}

\item{show_only_with_section}{render the output only if the section is visible.}
}
\value{
the path to the prerendered HTML file
}
\description{
Precompile HTML files
}
