% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_text_regexp.R
\name{question_is_correct.text_regexp}
\alias{question_is_correct.text_regexp}
\title{Check the answer to a text question against a regular expression}
\usage{
\method{question_is_correct}{text_regexp}(question, value, ...)
}
\description{
Check the answer to a text question against a regular expression
}
