% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_text_regexp.R
\name{question_is_valid.text_regexp}
\alias{question_is_valid.text_regexp}
\title{Validate a text question which checks the answer against a regular expression}
\usage{
\method{question_is_valid}{text_regexp}(question, value, ...)
}
\description{
Validate a text question which checks the answer against a regular expression
}
