% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_pool.R
\name{question_pool}
\alias{question_pool}
\title{Create a question with answers randomly taken from the pool of answers.}
\usage{
question_pool(
  title,
  ...,
  nr_answers,
  random_answer_order = TRUE,
  container_class = NULL,
  post_rendered = NULL,
  title_container = h5,
  show_only_with_section = TRUE
)
}
\arguments{
\item{title}{question title.}

\item{...}{answer options.}

\item{nr_answers}{number of incorrect answers to display. By default one correct answer and all incorrect answers
are shown.}

\item{random_answer_order}{should the order of answers be randomized?}

\item{container_class}{additional HTML classes to add to the container.}

\item{post_rendered}{optional function to post-process rendered text (e.g., to replace placeholders). The function
is called with a single string marked as HTML.}

\item{show_only_with_section}{render the output only if the section is visible.}
}
\description{
Create a question with answers randomly taken from the pool of answers.
}
