% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_text_regexp.R
\name{question_ui_initialize.text_regexp}
\alias{question_ui_initialize.text_regexp}
\title{Initialize a text question which checks the answer against a regular expression}
\usage{
\method{question_ui_initialize}{text_regexp}(question, value, ...)
}
\description{
Initialize a text question which checks the answer against a regular expression
}
