% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{read_session_cookie}
\alias{read_session_cookie}
\title{Read or Create the Session Cookie}
\usage{
read_session_cookie(
  appid,
  domain = shiny::getDefaultReactiveDomain(),
  param_name = "userid",
  session_length = 0.125
)
}
\arguments{
\item{appid}{Application id. All applications with the same ID share the same session.}

\item{domain}{Shiny reactive domain.}

\item{param_name}{URL query string parameter name to extract the user id from.}

\item{session_length}{Duration for how long the session is to be active (in days).}
}
\value{
the user id as given in the query string or as saved in the user session.
}
\description{
The user session contains the user's id and is stored in the client's browser as cookie.
This function should not be called directly by itself, but rather through the setup function \link{setup_user_session}.
}
\details{
For the session logic to work, the shiny app needs to have \link{shinyjs} loaded and the \emph{session.js} file included from
\code{system.file('inst', 'js', 'session.js', package = 'stat305templates')}.
See the templates as provided by this package how this can be accomplished.
}
