% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_format.R
\name{render_lab_answers}
\alias{render_lab_answers}
\title{Render and Validate Lab Answers}
\usage{
render_lab_answers(filename, output_dir = NULL)
}
\arguments{
\item{filename}{path to the lab answer file.}

\item{output_dir}{the output directory for the rendered answer file. If \code{NULL}, same as the directory of the input
file.}
}
\description{
Renders and validates lab answers generated by and downloaded from a lab tutorial.
}
