% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{render_markdown_as_html}
\alias{render_markdown_as_html}
\title{Render markdown text as HTML}
\usage{
render_markdown_as_html(text)
}
\arguments{
\item{text}{the markdown text to be parsed as HTML}
}
\value{
a HTML container with the parsed markdown.
}
\description{
Translate markdown into HTML elements.
}
