% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rmd.R
\name{render_rmd_server}
\alias{render_rmd_server}
\title{Render the given R Markdown file \emph{on the server}}
\usage{
render_rmd_server(
  file,
  data = list(),
  runtime = "static",
  output_format,
  quiet = TRUE,
  show_only_with_section = TRUE
)
}
\arguments{
\item{file}{path to the RMarkdown file}

\item{data}{a list of data to be made available to the Rmd file during rendering. Quoted items will be evaluated
on the server before knitting the Rmd file. Quoted expressions are evaluated in the \strong{server context}.}

\item{runtime, output_format, quiet}{arguments passed on to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}

\item{show_only_with_section}{render the output only if the section is visible.}
}
\value{
the UI output element
}
\description{
Render the given R Markdown file \emph{on the server}
}
