% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_session_data}
\alias{set_session_data}
\title{Get the default session data}
\usage{
set_session_data(
  what,
  value,
  session = getDefaultReactiveDomain(),
  asis = FALSE
)
}
\description{
Get the default session data
}
