% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential-sections.R
\name{show_section}
\alias{show_section}
\title{Show a specific section.}
\usage{
show_section(section, hide_others = TRUE)
}
\arguments{
\item{section}{section to show.}

\item{hide_others}{hide all other sections.}

\item{session}{shiny reactive domain.}
}
\description{
Show a specific section.
}
