% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{start_user_session}
\alias{start_user_session}
\title{Start or Resume a User Session}
\usage{
start_user_session(
  appid,
  userid_param_name = "userid",
  default_userid = "unknown",
  domain = getDefaultReactiveDomain(),
  session_length = 0.125
)
}
\arguments{
\item{appid}{application id. All applications with the same ID share the same session (e.g., to store the user id).}

\item{userid_param_name}{query string parameter name for the user id.}

\item{default_userid}{fallback user id if it can not be determined from the query string or read/written from/to a
cookie.}

\item{domain}{Shiny reactive domain.}

\item{session_length}{how long (in days) should the user session last?}
}
\description{
Start or Resume a User Session
}
