% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_question.R
\name{text_question}
\alias{text_question}
\title{Add a text question.}
\usage{
text_question(
  question,
  label = NULL,
  textarea = TRUE,
  title_container = h5,
  width = NULL,
  height = NULL,
  placeholder = NULL
)
}
\arguments{
\item{question}{the question text. Markdown is supported.}

\item{label}{an optional label for the input element.}

\item{textarea}{if \code{TRUE} (default) uses a HTML \emph{textarea} element for the input.
If \code{FALSE}, uses an \emph{input} element.}

\item{width}{the width of the input, e.g., '300px' or '100\%'; see \code{\link[shiny:validateCssUnit]{shiny::validateCssUnit()}}.}

\item{height}{the height of the input, e.g., '300px' or '100\%'; see \code{\link[shiny:validateCssUnit]{shiny::validateCssUnit()}}.
Only has an effect if \code{textarea=TRUE}.}

\item{placeholder}{A character string giving the user a hint as to what can be entered into the control.
Internet Explorer 8 and 9 do not support this option.}
}
\description{
Add a text question.
}
