% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-formats.R
\name{worksheet}
\alias{worksheet}
\title{Worksheet Format}
\usage{
worksheet(...)
}
\description{
Displays the content in large font and transforms 3rd-level sections into questions which are by default
collapsed.
}
