% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_safe.R
\name{enable_safe_eval}
\alias{enable_safe_eval}
\title{Enable Safe Evaluation of User Code for \code{learnr} Exercises}
\usage{
enable_safe_eval(
  max_fsize = 1L,
  max_address_space = 2048L,
  max_data = 256L,
  priority = 20L,
  rlimits,
  allow_env,
  default_allow_env = c("^SHELL$", "^USER$", "^LANG$", "^LC_CTYPE$", "^HOME$",
    "^DYLD_FALLBACK_LIBRARY_PATH$", "^R_.*")
)
}
\arguments{
\item{max_fsize}{maximum size of created files in MiB.}

\item{max_address_space}{maximum size of the address space in MiB.}

\item{max_data}{maximum size of the data memory in MiB.}

\item{priority}{priority of the evaluating process.}

\item{rlimits}{specify custom rlimitis.}

\item{allow_env}{vector of regular expressions of allowed environment variables (i.e., the matching variables will
be retained).}

\item{default_allow_env}{vector of regular expressions of environment variables which should be allowed by default.
Only change if you know what you are doing, as most of them are fundamental to the R process
to work.}
}
\description{
Enable Safe Evaluation of User Code for \code{learnr} Exercises
}
