% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential-sections.R
\name{init_sections}
\alias{init_sections}
\title{Initialize sections for this session}
\usage{
init_sections(sections, hide = TRUE, session = getDefaultReactiveDomain())
}
\arguments{
\item{sections}{the sections object}

\item{hide}{should sections be hidden by default.}
}
\description{
Initialize sections for this session
}
