% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_clicker.R, R/question_pool.R,
%   R/render_rmd.R, R/sequential-sections.R
\name{knit_print.clicker_question}
\alias{knit_print.clicker_question}
\alias{knit_print.question_pool}
\alias{knit_print.rendered_rmd}
\alias{knit_print.sequential_section_begin}
\alias{knit_print.sequential_section_end}
\title{Knitr clicker question print methods}
\usage{
\method{knit_print}{clicker_question}(x, ...)

\method{knit_print}{question_pool}(x, ...)

\method{knit_print}{rendered_rmd}(x, ...)

\method{knit_print}{sequential_section_begin}(x, ...)

\method{knit_print}{sequential_section_end}(x, ...)
}
\arguments{
\item{x}{An R object to be printed}

\item{...}{Additional arguments passed to the S3 method. Currently ignored,
except two optional arguments \code{options} and \code{inline}; see
the references below.}
}
\description{
\link[knitr:knit_print]{knitr::knit_print} method for \link{clicker_question}.

\link[knitr:knit_print]{knitr::knit_print} method for \link{question_pool}.

\link[knitr:knit_print]{knitr::knit_print} method for \link{question_pool}.
}
