% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_format.R
\name{lab_name_input}
\alias{lab_name_input}
\title{Add a text input for the user's name}
\usage{
lab_name_input(title = "What is your name?", label = "Your name")
}
\arguments{
\item{title}{title for the panel containing the text input.}

\item{label}{label of the text input.}
}
\description{
Add a text input for the user's name
}
