% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_pool.R
\name{question_pool}
\alias{question_pool}
\title{Create a question with answers randomly taken from the pool of answers.}
\usage{
question_pool(
  title,
  ...,
  nr_answers = 5,
  random_answer_order = TRUE,
  container_class = NULL,
  mc = FALSE,
  post_rendered = NULL,
  title_container = h5,
  correct_label = 12,
  show_only_with_section = TRUE
)
}
\arguments{
\item{title}{question title.}

\item{...}{answer options.}

\item{nr_answers}{maximum number of answers to display. At least one correct answer will always be shown.}

\item{random_answer_order}{should the order of answers be randomized?}

\item{container_class}{additional HTML classes to add to the container.}

\item{mc}{show as multiple choice.}

\item{post_rendered}{optional function to post-process rendered text (e.g., to replace placeholders). The function
is called with a single string marked as HTML and evaluated in the \strong{server context}.}

\item{correct_label}{if not \code{NULL}, change the answer option labels to hex numbers with the right-most digit
being equal to the given number (by default 0xa). The length of the hex number is determined by the number of
answer options.}

\item{show_only_with_section}{render the output only if the section is visible.}
}
\description{
Create a question with answers randomly taken from the pool of answers.
}
