% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_format.R
\name{render_lab_answers}
\alias{render_lab_answers}
\title{Render and Validate Lab Answers}
\usage{
render_lab_answers(filename, output_dir = NULL, zip_archive)
}
\arguments{
\item{filename}{path to the lab answer file.}

\item{output_dir}{the output directory for the rendered answer file. If \code{NULL}, the output directory depends on the
type of the file. If it is a markdown file, the output will be saved in the directory of the input.
If it is a ZIP archive, all the files in the ZIP file will be rendered in the current working directory.}

\item{zip_archive}{is the given file a ZIP archive? If yes, process all files in the archive.
By default, files ending in ".zip" are treated as ZIP archives.}
}
\description{
Renders and validates lab answers generated by and downloaded from a lab tutorial.
}
