% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_session_seed}
\alias{set_session_seed}
\title{Set the "master" seed for the package.}
\usage{
set_session_seed(seed, session)
}
\arguments{
\item{seed}{seed.}

\item{session}{the shiny reactive domain.}
}
\description{
All subsequent randomizations by the \code{stat305templates} package are based on this master seed.
}
