% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_format.R
\name{submit_lab_btn}
\alias{submit_lab_btn}
\title{Add a button to submit the lab.}
\usage{
submit_lab_btn(
  label = "Download answers",
  filename = "lab_answers.md",
  require_validation = TRUE,
  render_html = FALSE
)
}
\arguments{
\item{label}{button label}

\item{filename}{name of the file offered to the user for download. The file extension is irrelevant.}

\item{require_validation}{create a markdown file with validation information.}

\item{render_html}{directly render the lab as HTML instead of outputting a markdown file.
If \code{TRUE}, validation metadata cannot be included.}
}
\description{
A name input needs to be placed \emph{before} the submit button using \code{\link[=lab_name_input]{lab_name_input()}}.
}
